refactor: ファイル構造の最適化とクリーンアップ
WorkoutFormのファイル構造を改善し、コードベースをクリーンアップ。
コロケーションの原則に基づき、関連ファイルを同一ディレクトリに集約。

主な変更:
1. ファイル構造の改善
   - WorkoutForm.jsx を WorkoutForm/index.jsx に移動
   - 関連コンポーネントを同一フォルダに集約
   - インポートパスの整理

2. 責任範囲の明確化
   - pages/WorkoutForm: ページレイアウト、タイトル管理
   - components/WorkoutForm: 純粋なUIロジック

3. コードクリーンアップ
   - 不要なコメント16箇所を削除
   - React.memoによるパフォーマンス最適化
   - 未使用インポートの削除

効果:
- モジュールの境界が明確化（フォルダ = 1コンポーネント）
- インポートパスの簡潔性を維持
- 将来的な拡張性（テスト、スタイル追加時も整理された状態）
